4th place in first hackathon
we won 4th place in INTER COLLEGIATE HACKATHON 2020 conducted by INFOSYS on Jan 8, 2020
due to some issues in laptop we not able to show our final demo, inspire of that due to voting by some master hacks who seen our project before we got this place. we are always thanking to them. As it was our first hackathon we got lot of experience and knowledge.
#T20hack #InfyDiaries #InfyChennaiDC